Sanitize metadata in chunks
Instead of loading all metadata into memory to perform transformation
and filtering associated with sanitizing logic, process metadata in
chunks that easily fit in memory.

To support this low-memory implementation, this commit changes the
approach to resolving duplicate records. We now make a first pass
through the metadata to store strain ids and corresponding database ids
for each record in a temporary file on disk. We make a second pass to
filter duplicate records, transform the remaining records, and stream
the results to disk.

This commit adds cram-based functional tests that were used for
test-driven development of the new low-memory functionality and to
confirm that the user interface remains stable.

One minor interface change (improvement) from this refactoring is the
creation of a separate file containing the list of all duplicate strains
when the user has requested an error on duplicates. This change allows
the user to bioinformatically address duplicates after the fact in a way
that printing the list of strains to stderr did not easily allow.