Move (nextstrain) alignment into separate profiles
This breaks each of our main nextstrain profiles into two profiles each - one to perform alignment (etc) and a second to run the phylogenetics. This is part of a wider effort to generate alignments as soon as new sequences are available.

Using GISAID as an example (also implemented for open, read the following but replace `gisaid` with `open` ):

The first profile `nextstrain_profiles/nextstrain-gisaid-preprocess` takes sequences and metadata and produces three files which we upload to S3,  `results/filtered_gisaid.fasta.xz`, `results/masked_gisaid.fasta.xz` and `results/aligned_gisaid.fasta.xz`.  While this is a separate profile, the  `builds.yaml` file is around a dozen lines of code.

The second profile, `nextstrain_profiles/nextstrain-gisaid` is relatively unchanged, except that we now start from `results/filtered_gisaid.fasta.xz` and thus should be much faster to run. Note that the `upload`  rule here will no longer upload the files which are now within the previous profile. Iit is still possible to start this workflow from (unaligned) sequences, however there should be no reason to do so.

Note that both profiles will run `rule sanitize_metadata`, as each depend on its output which is not uploaded as an intermediate file.

None of these changes should affect non-nextstrain-core builds / profiles.